requirements.txt
Specify dependencies (none required for this extension):
```plaintext
ulauncher-api
